move currencyInputPanel out
